+++ pybamm/expression_tree/binary_operators.py
@@ -162,12 +162,6 @@ class BinaryOperator(pybamm.Symbol):
         """ Calculate the jacobian of a binary operator. """
         raise NotImplementedError
 
-    def _binary_simplify(self, new_left, new_right):
-        """ Simplify a binary operator """
-        return pybamm.simplify_if_constant(
-            self._binary_new_copy(new_left, new_right), clear_domains=False
-        )
-
     def _binary_evaluate(self, left, right):
         """ Perform binary operation on nodes 'left' and 'right'. """
         raise NotImplementedError
@@ -226,10 +220,6 @@ class Power(BinaryOperator):
         with np.errstate(invalid="ignore"):
             return left ** right
 
-    def _binary_simplify(self, new_left, new_right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplified_power(new_left, new_right)
-
 
 class Addition(BinaryOperator):
     """A node in the expression tree representing an addition operator
@@ -252,10 +242,6 @@ class Addition(BinaryOperator):
     def _binary_evaluate(self, left, right):
         """ See :meth:`pybamm.BinaryOperator._binary_evaluate()`. """
         return left + right
-
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_addition_subtraction(self.__class__, left, right)
 
 
 class Subtraction(BinaryOperator):
@@ -280,12 +266,6 @@ class Subtraction(BinaryOperator):
     def _binary_evaluate(self, left, right):
         """ See :meth:`pybamm.BinaryOperator._binary_evaluate()`. """
         return left - right
-
-    def _binary_simplify(self, left, right):
-        """
-        See :meth:`pybamm.BinaryOperator._binary_simplify()`.
-        """
-        return pybamm.simplify_addition_subtraction(self.__class__, left, right)
 
 
 class Multiplication(BinaryOperator):
@@ -332,10 +312,6 @@ class Multiplication(BinaryOperator):
         else:
             return left * right
 
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_multiplication_division(self.__class__, left, right)
-
 
 class MatrixMultiplication(BinaryOperator):
     """A node in the expression tree representing a matrix multiplication operator
@@ -377,10 +353,6 @@ class MatrixMultiplication(BinaryOperator):
     def _binary_evaluate(self, left, right):
         """ See :meth:`pybamm.BinaryOperator._binary_evaluate()`. """
         return left @ right
-
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_multiplication_division(self.__class__, left, right)
 
 
 class Division(BinaryOperator):
@@ -424,10 +396,6 @@ class Division(BinaryOperator):
                     return left * np.inf
             else:
                 return left / right
-
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_multiplication_division(self.__class__, left, right)
 
 
 class Inner(BinaryOperator):
@@ -485,10 +453,6 @@ class Inner(BinaryOperator):
     def _binary_new_copy(self, left, right):
         """ See :meth:`pybamm.BinaryOperator._binary_new_copy()`. """
         return pybamm.inner(left, right)
-
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_multiplication_division(self.__class__, left, right)
 
     def _evaluates_on_edges(self, dimension):
         """ See :meth:`pybamm.Symbol._evaluates_on_edges()`. """