+++ pybamm/expression_tree/concatenations.py
@@ -94,12 +94,6 @@ class Concatenation(pybamm.Symbol):
         """ Calculate the jacobian of a concatenation """
         return NotImplementedError
 
-    def _concatenation_simplify(self, children):
-        """ See :meth:`pybamm.Symbol.simplify()`. """
-        new_symbol = self.__class__(*children)
-        new_symbol.clear_domains()
-        return new_symbol
-
     def _evaluate_for_shape(self):
         """ See :meth:`pybamm.Symbol.evaluate_for_shape` """
         if len(self.children) == 0:
@@ -154,12 +148,6 @@ class NumpyConcatenation(Concatenation):
             return pybamm.Scalar(0)
         else:
             return SparseStack(*children_jacs)
-
-    def _concatenation_simplify(self, children):
-        """ See :meth:`pybamm.Concatenation._concatenation_simplify()`. """
-        new_symbol = simplified_numpy_concatenation(*children)
-        new_symbol.clear_domains()
-        return new_symbol
 
 
 class DomainConcatenation(Concatenation):
@@ -306,17 +294,6 @@ class DomainConcatenation(Concatenation):
         new_symbol = simplified_domain_concatenation(
             children, self.full_mesh, copy_this=self
         )
-        return new_symbol
-
-    def _concatenation_simplify(self, children):
-        """ See :meth:`pybamm.Concatenation._concatenation_simplify()`. """
-        new_symbol = simplified_domain_concatenation(
-            children, self.full_mesh, copy_this=self
-        )
-        # TODO: this should not be needed, but somehow we are still getting domains in
-        # the simplified children
-        new_symbol.clear_domains()
-
         return new_symbol